+++ Garage/src/com/option/Climatisation.java
@@ -7,7 +7,7 @@ public class Climatisation implements Option{
 	
 	// methode getPrix
 	@Override
-	public float getPrix() {
+	public double getPrix() {
 		return prix;
 	}
